requirements.txt
boto3
pandas
jinja2
psycopg2-binary
pandas